refactor: 엔티티 자료형에 맞게 Long -> Integer로 변경

+++ src/main/java/com/jandi/band_backend/club/controller/ClubEventController.java
@@ -36,7 +36,7 @@ public class ClubEventController {
     @GetMapping("/events/{eventId}")
     public ResponseEntity<CommonResponse<ClubEventRespDTO>> getClubEventDetail(
             @PathVariable Integer clubId,
-            @PathVariable Long eventId,
+            @PathVariable Integer eventId,
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
         Integer userId = userDetails.getUserId();
@@ -65,7 +65,7 @@ public class ClubEventController {
     @DeleteMapping("/events/{eventId}")
     public ResponseEntity<CommonResponse<Void>> deleteClubEvent(
             @PathVariable Integer clubId,
-            @PathVariable Long eventId,
+            @PathVariable Integer eventId,
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
         Integer userId = userDetails.getUserId();
